Move `LocationPreview` into a separate module
Working towards simplifying the `RoomScreen` into a less-complex megalith.
