+++ src/home/room_screen.rs
@@ -1,7 +1,7 @@
-//! A room screen is the UI page that displays a single Room's timeline of events/messages
+//! A room screen is the UI view that displays a single Room's timeline of events/messages
 //! along with a message input bar at the bottom.
 
-use std::{borrow::Cow, collections::BTreeMap, ops::{DerefMut, Range}, sync::{Arc, Mutex}, time::SystemTime};
+use std::{borrow::Cow, collections::BTreeMap, ops::{DerefMut, Range}, sync::{Arc, Mutex}};
 
 use bytesize::ByteSize;
 use imbl::Vector;
@@ -17,10 +17,9 @@ use matrix_sdk::{room::RoomMember, ruma::{
 use matrix_sdk_ui::timeline::{
     self, EventTimelineItem, InReplyToDetails, MemberProfileChange, RepliedToInfo, RoomMembershipChange, TimelineDetails, TimelineEventItemId, TimelineItem, TimelineItemContent, TimelineItemKind, VirtualTimelineItem
 };
-use robius_location::Coordinates;
 
 use crate::{
-    avatar_cache, event_preview::{body_of_timeline_item, text_preview_of_member_profile_change, text_preview_of_other_state, text_preview_of_redacted_message, text_preview_of_room_membership_change, text_preview_of_timeline_item}, home::loading_pane::{LoadingPaneState, LoadingPaneWidgetExt}, location::{get_latest_location, init_location_subscriber, request_location_update, LocationAction, LocationRequest, LocationUpdate}, media_cache::{MediaCache, MediaCacheEntry}, profile::{
+    avatar_cache, event_preview::{body_of_timeline_item, text_preview_of_member_profile_change, text_preview_of_other_state, text_preview_of_redacted_message, text_preview_of_room_membership_change, text_preview_of_timeline_item}, home::loading_pane::{LoadingPaneState, LoadingPaneWidgetExt}, location::init_location_subscriber, media_cache::{MediaCache, MediaCacheEntry}, profile::{
         user_profile::{AvatarState, ShowUserProfileAction, UserProfile, UserProfileAndRoomId, UserProfilePaneInfo, UserProfileSlidingPaneRef, UserProfileSlidingPaneWidgetExt},
         user_profile_cache,
     }, shared::{
@@ -34,7 +33,7 @@ use crate::shared::mentionable_text_input::MentionableTextInputWidgetRefExt;
 
 use rangemap::RangeSet;
 
-use super::{editing_pane::{EditingPaneAction, EditingPaneWidgetExt}, event_reaction_list::ReactionData, loading_pane::LoadingPaneRef, new_message_context_menu::{MessageAbilities, MessageDetails}, room_read_receipt::{self, populate_read_receipts, MAX_VISIBLE_AVATARS_IN_READ_RECEIPT}};
+use super::{editing_pane::{EditingPaneAction, EditingPaneWidgetExt}, event_reaction_list::ReactionData, loading_pane::LoadingPaneRef, location_preview::LocationPreviewWidgetExt, new_message_context_menu::{MessageAbilities, MessageDetails}, room_read_receipt::{self, populate_read_receipts, MAX_VISIBLE_AVATARS_IN_READ_RECEIPT}};
 
 const GEO_URI_SCHEME: &str = "geo:";
 
@@ -52,15 +51,17 @@ live_design! {
     use crate::shared::text_or_image::TextOrImage;
     use crate::shared::html_or_plaintext::*;
     use crate::shared::icon_button::*;
-    use crate::home::room_read_receipt::*;
-    use crate::profile::user_profile::UserProfileSlidingPane;
     use crate::shared::typing_animation::TypingAnimation;
     use crate::shared::icon_button::*;
     use crate::shared::jump_to_bottom_button::*;
-    use crate::home::loading_pane::*;
-    use crate::home::event_reaction_list::*;
+    use crate::profile::user_profile::UserProfileSlidingPane;
     use crate::home::editing_pane::*;
+    use crate::home::event_reaction_list::*;
+    use crate::home::loading_pane::*;
+    use crate::home::location_preview::*;
     use crate::room::room_input_bar::*;
+    use crate::room::room_input_bar::*;
+    use crate::home::room_read_receipt::*;
 
     IMG_DEFAULT_AVATAR = dep("crate://self/resources/img/default_avatar.png")
 
@@ -197,27 +198,6 @@ live_design! {
         }
     }
 
-    // An optional view used to show reactions beneath a message.
-    MessageAnnotations = <View> {
-        visible: false,
-        width: Fill,
-        height: Fit,
-        padding: {top: 5.0}
-
-        html_content = <MessageHtml> {
-            width: Fill,
-            height: Fit,
-            padding: { bottom: 5.0, top: 0.0 },
-            font_size: 10.5,
-            font_color: (REACTION_TEXT_COLOR),
-            draw_normal:      { color: (REACTION_TEXT_COLOR) },
-            draw_italic:      { color: (REACTION_TEXT_COLOR) },
-            draw_bold:        { color: (REACTION_TEXT_COLOR) },
-            draw_bold_italic: { color: (REACTION_TEXT_COLOR) },
-            draw_fixed:       { color: (REACTION_TEXT_COLOR) },
-            body: ""
-        }
-    }
 
     // An empty view that takes up no space in the portal list.
     Empty = <View> { }
@@ -616,91 +596,6 @@ live_design! {
         jump_to_bottom = <JumpToBottomButton> { }
     }
 
-    LocationPreview = {{LocationPreview}} {
-        visible: false
-        width: Fill
-        height: Fit
-        flow: Down
-        padding: {left: 12.0, top: 12.0, bottom: 12.0, right: 10.0}
-        spacing: 15
-
-        show_bg: true,
-        draw_bg: {
-            color: #xF0F5FF,
-        }
-
-        <Label> {
-            width: Fill,
-            height: Fit,
-            draw_text: {
-                wrap: Word,
-                color: (MESSAGE_TEXT_COLOR),
-                text_style: <MESSAGE_TEXT_STYLE>{ font_size: 10.0 },
-            }
-            text: "Send your location to this room?"
-        }
-
-        location_label = <Label> {
-            width: Fill,
-            height: Fit,
-            align: {x: 0.0, y: 0.5},
-            padding: {left: 5.0}
-            draw_text: {
-                wrap: Word,
-                color: (MESSAGE_TEXT_COLOR),
-                text_style: <MESSAGE_TEXT_STYLE>{},
-            }
-            text: "Fetching current location..."
-        }
-
-        <View> {
-            width: Fill, height: Fit
-            flow: Right,
-            align: {x: 0.0, y: 0.5}
-            spacing: 15
-
-            cancel_location_button = <RobrixIconButton> {
-                align: {x: 0.5, y: 0.5}
-                padding: 15,
-                draw_icon: {
-                    svg_file: (ICON_BLOCK_USER)
-                    color: (COLOR_DANGER_RED),
-                }
-                icon_walk: {width: 16, height: 16, margin: {left: -2, right: -1, top: -1} }
-
-                draw_bg: {
-                    border_color: (COLOR_DANGER_RED),
-                    color: #fff0f0 // light red
-                }
-                text: "Cancel"
-                draw_text:{
-                    color: (COLOR_DANGER_RED),
-                }
-            }
-
-            send_location_button = <RobrixIconButton> {
-                // disabled by default; will be enabled upon receiving valid location update.
-                enabled: false,
-                align: {x: 0.5, y: 0.5}
-                padding: 15,
-                draw_icon: {
-                    svg_file: (ICON_SEND)
-                    color: (COLOR_ACCEPT_GREEN),
-                }
-                icon_walk: {width: 16, height: 16, margin: {left: -2, right: -1} }
-
-                draw_bg: {
-                    border_color: (COLOR_ACCEPT_GREEN),
-                    color: #f0fff0 // light green
-                }
-                text: "Yes"
-                draw_text:{
-                    color: (COLOR_ACCEPT_GREEN),
-                }
-            }
-        }
-    }
-
 
     pub RoomScreen = {{RoomScreen}} {
         width: Fill, height: Fill,
@@ -869,7 +764,7 @@ live_design! {
 
 
             /*
-             * This is broken currently, so I'm disabling it.
+             * TODO: add the action bar back in as a series of floating buttons.
              *
             message_action_bar_popup = <PopupNotification> {
                 align: {x: 0.0, y: 0.0}
@@ -4073,112 +3968,6 @@ fn get_profile_display_name(event_tl_item: &EventTimelineItem) -> Option<String>
         profile.display_name.clone()
     } else {
         None
-    }
-}
-
-/// A simple deref wrapper around the `LocationPreview` widget that enables us to handle actions on it.
-#[derive(Live, LiveHook, Widget)]
-struct LocationPreview {
-    #[deref] view: View,
-    #[rust] coords: Option<Result<Coordinates, robius_location::Error>>,
-    #[rust] timestamp: Option<SystemTime>,
-}
-impl Widget for LocationPreview {
-    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
-        let mut needs_redraw = false;
-        if let Event::Actions(actions) = event {
-            for action in actions {
-                match action.downcast_ref() {
-                    Some(LocationAction::Update(LocationUpdate { coordinates, time })) => {
-                        self.coords = Some(Ok(*coordinates));
-                        self.timestamp = *time;
-                        self.button(id!(send_location_button)).set_enabled(cx, true);
-                        needs_redraw = true;
-                    }
-                    Some(LocationAction::Error(e)) => {
-                        self.coords = Some(Err(*e));
-                        self.timestamp = None;
-                        self.button(id!(send_location_button)).set_enabled(cx, false);
-                        needs_redraw = true;
-                    }
-                    _ => { }
-                }
-            }
-
-            // NOTE: the send location button click event is handled
-            //       in the RoomScreen handle_event function.
-
-            // Handle the cancel location button being clicked.
-            if self.button(id!(cancel_location_button)).clicked(actions) {
-                self.clear();
-                needs_redraw = true;
-            }
-        }
-
-        if needs_redraw {
-            self.redraw(cx);
-        }
-
-        self.view.handle_event(cx, event, scope);
-    }
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        let text = match self.coords {
-            Some(Ok(c)) => {
-                // if let Some(st) = self.timestamp {
-                //     format!("Current location: {:.6},{:.6}\n   Timestamp: {:?}", c.latitude, c.longitude, st)
-                // } else {
-                    format!("Current location: {:.6},{:.6}", c.latitude, c.longitude)
-                // }
-            }
-            Some(Err(e)) => format!("Error getting location: {e:?}"),
-            None => String::from("Current location is not yet available."),
-        };
-        self.label(id!(location_label)).set_text(cx, &text);
-        self.view.draw_walk(cx, scope, walk)
-    }
-}
-
-
-impl LocationPreview {
-    fn show(&mut self) {
-        request_location_update(LocationRequest::UpdateOnce);
-        if let Some(loc) = get_latest_location() {
-            self.coords = Some(Ok(loc.coordinates));
-            self.timestamp = loc.time;
-        }
-        self.visible = true;
-    }
-
-    fn clear(&mut self) {
-        self.coords = None;
-        self.timestamp = None;
-        self.visible = false;
-    }
-
-    pub fn get_current_data(&self) -> Option<(Coordinates, Option<SystemTime>)> {
-        self.coords
-            .as_ref()
-            .and_then(|res| res.ok())
-            .map(|c| (c, self.timestamp))
-    }
-}
-
-impl LocationPreviewRef {
-    pub fn show(&self) {
-        if let Some(mut inner) = self.borrow_mut() {
-            inner.show();
-        }
-    }
-
-    pub fn clear(&self) {
-        if let Some(mut inner) = self.borrow_mut() {
-            inner.clear();
-        }
-    }
-
-    pub fn get_current_data(&self) -> Option<(Coordinates, Option<SystemTime>)> {
-        self.borrow().and_then(|inner| inner.get_current_data())
     }
 }
